+++ app/src/main/java/com/app/diploma/domain/CryptoRepository.kt
@@ -6,6 +6,6 @@ interface CryptoRepository {
 
     suspend fun getAllCurrencies(): Result<List<Currency>>
 
-    suspend fun getRateByCurrency(currency: Currency): Result<Long>
+    suspend fun getRateByCurrency(currency: Currency): Result<Double>
 
 }
